Move adapter + rollout into per-backend module paths (zw0610 #2/#3/#4)
Addresses @zw0610's structural review on verl-project/verl#6422:

- #2: verl/trainer/ppo/arctic_rl_client.py
       -> verl/workers/remote_client/arctic_rl_client.py
- #3: verl/workers/rollout/arctic_rollout/
       -> verl/workers/rollout/remote_rollout/arctic_rollout/
- #4: TODO(@zw0610) on RolloutReplicaRegistry registrations in
       verl/workers/rollout/replica.py:360 about lazy-init each option to
       avoid pulling in vLLM + SGLang + TRT-LLM transitive deps eagerly.

Pure code-motion + a TODO; no behavior change. Imports inside the moved
files and in workers/rollout/replica.py updated.
