+++ verl/workers/rollout/remote_rollout/arctic_rollout/arctic_rollout.py
@@ -4,7 +4,7 @@ from verl.workers.rollout.vllm_rollout.vllm_async_server import vLLMHttpServer
 
 import argparse
 from typing import Any, Optional
-from verl.trainer.ppo.arctic_rl_client import ArcticRLClientWrapper
+from verl.workers.remote_client.arctic_rl_client import ArcticRLClientWrapper
 from collections.abc import AsyncGenerator
 
 import ray
