+++ verl/workers/rollout/replica.py
@@ -349,10 +349,17 @@ def _load_trtllm():
     return TRTLLMReplica
 
 def _load_arctic():
-    from verl.workers.rollout.arctic_rollout.arctic_rollout import ArcticReplica
-    
+    from verl.workers.rollout.remote_rollout.arctic_rollout.arctic_rollout import ArcticReplica
+
     return ArcticReplica
 
+
+# TODO(@zw0610): lazy-init each option for `RolloutReplicaRegistry`. For most
+# cases there will be only 1 rollout backend used, but importing all of them
+# eagerly could surface dependency conflicts (e.g. pulling in vLLM + SGLang +
+# TRT-LLM simultaneously). The `_load_*` callables above already defer imports
+# to call-time; the registration calls below could be moved into a per-backend
+# `register_<name>()` helper that the user calls explicitly at startup.
 # Register built-in types
 RolloutReplicaRegistry.register("vllm", _load_vllm)
 RolloutReplicaRegistry.register("sglang", _load_sglang)
